[NL Next] Break down fulfillment into modules by classification type
